+++ sovtimer/views.py
@@ -11,12 +11,9 @@ from django.http import JsonResponse
 from django.shortcuts import render
 from django.utils.translation import gettext_lazy as _
 
-from eveuniverse.models import EveSolarSystem
-
 from sovtimer import __title__
 from sovtimer.app_settings import avoid_cdn
 from sovtimer.models import AaSovtimerCampaigns, AaSovtimerStructures
-from sovtimer.providers import esi
 from sovtimer.utils import LoggerAddTag
 
 from allianceauth.services.hooks import get_extension_logger
@@ -32,6 +29,8 @@ MAP_EVENT_TO_TYPE = {
     "station_defense": "Station",
     "station_freeport": "Freeport",
 }
+
+CAMPAIGN_TREND_CACHE_TIME = 30
 
 
 @login_required
@@ -63,16 +62,9 @@ def dashboard_data(request) -> JsonResponse:
     sovereignty_structures = AaSovtimerStructures.objects.all()
 
     if sovereignty_campaigns and sovereignty_structures:
-        # sovereignty_structures_esi = (
-        #     esi.client.Sovereignty.get_sovereignty_structures().results()
-        # )
-
         for campaign in sovereignty_campaigns:
-            alliance_esi = esi.client.Alliance.get_alliances_alliance_id(
-                alliance_id=campaign.defender_id
-            ).results()
-
-            defender_name = alliance_esi["name"]
+            # defender name
+            defender_name = campaign.defender.name
             defender_name_html = (
                 '<a href="https://evemaps.dotlan.net/search?q={defender_name}" '
                 'target="_blank" rel="noopener noreferer">{defender_name}</a>'.format(
@@ -80,9 +72,8 @@ def dashboard_data(request) -> JsonResponse:
                 )
             )
 
-            eve_solar_system = EveSolarSystem.objects.get(id=campaign.solar_system_id)
-
-            solar_system_name = eve_solar_system.name
+            # solar system
+            solar_system_name = campaign.solar_system.name
             solar_system_name_html = (
                 '<a href="https://evemaps.dotlan.net/search?q={solar_system_name}" '
                 'target="_blank" rel="noopener noreferer">{solar_system_name}</a>'.format(
@@ -90,7 +81,8 @@ def dashboard_data(request) -> JsonResponse:
                 )
             )
 
-            constellation_name = eve_solar_system.eve_constellation.name
+            # constellation
+            constellation_name = campaign.solar_system.eve_constellation.name
             constellation_name_html = (
                 '<a href="https://evemaps.dotlan.net/search?q={constellation_name}" '
                 'target="_blank" rel="noopener noreferer">{constellation_name}</a>'.format(
@@ -98,7 +90,8 @@ def dashboard_data(request) -> JsonResponse:
                 )
             )
 
-            region_name = eve_solar_system.eve_constellation.eve_region.name
+            # region
+            region_name = campaign.solar_system.eve_constellation.eve_region.name
             region_name_html = (
                 '<a href="https://evemaps.dotlan.net/search?q={region_name}" '
                 'target="_blank" rel="noopener noreferer">{region_name}</a>'.format(
@@ -106,6 +99,7 @@ def dashboard_data(request) -> JsonResponse:
                 )
             )
 
+            # activity defense multiplier
             structure_adm = 1
             for structure in sovereignty_structures:
                 if (
@@ -113,57 +107,91 @@ def dashboard_data(request) -> JsonResponse:
                 ) and structure.vulnerability_occupancy_level:
                     structure_adm = structure.vulnerability_occupancy_level
 
+            # start time
             start_time = campaign.start_time.replace(tzinfo=None)
 
+            # remaining time field
             remaining_time_in_seconds = dt.timedelta(
                 seconds=(start_time.timestamp() - dt.datetime.now().timestamp())
             ).total_seconds()
 
-            attackers_score = campaign.attackers_score
-            attackers_score_percent = "{:.0f}%".format(attackers_score * 100)
-            if (attackers_score * 100) < 10:
-                attackers_score_percent = "0" + attackers_score_percent
+            # campaign progress field
+            campaign_pogress_current = campaign.progress_current
+            campaign_progress_previous = campaign.progress_previous
+            campaign_pogress_current_percentage = "{:.0f}%".format(
+                campaign_pogress_current * 100
+            )
 
-            defender_score = campaign.defender_score
-            defender_score_percent = "{:.0f}%".format(defender_score * 100)
-            if (defender_score * 100) < 10:
-                defender_score_percent = "0" + defender_score_percent
+            campaign_progress_html = campaign_pogress_current_percentage
+            if (campaign_pogress_current * 100) < 10:
+                campaign_progress_html = "0" + campaign_progress_html
 
             active_campaign = _("No")
             if remaining_time_in_seconds < 0:
                 active_campaign = _("Yes")
+
+                campaign_progress_icon = (
+                    '<i class="material-icons aa-sovtimer-trend aa-sovtimer-trend-flat" '
+                    'title="{title_text}">trending_flat</i>'.format(
+                        title_text=_("Neither side has made any progress yet")
+                    )
+                )
+
+                if campaign_progress_previous < campaign_pogress_current:
+                    campaign_progress_icon = (
+                        '<i class="material-icons aa-sovtimer-trend aa-sovtimer-trend-up" '
+                        'title="{title_text}">trending_up</i>'.format(
+                            title_text=_("Defenders making progress")
+                        )
+                    )
+
+                if campaign_progress_previous > campaign_pogress_current:
+                    campaign_progress_icon = (
+                        '<i class="material-icons aa-sovtimer-trend aa-sovtimer-trend-down" '
+                        'title="{title_text}">trending_down</i>'.format(
+                            title_text=_("Attackers making progress")
+                        )
+                    )
+
                 constellation_killboard_link = (
                     '<a href="https://zkillboard.com/constellation/{constellation_id}/" '
                     'target="_blank" rel="noopener noreferer" '
                     'class="aa-sov-timer-zkb-icon">{zkb_icon}</a>'.format(
-                        constellation_id=campaign.constellation_id,
+                        constellation_id=campaign.solar_system.eve_constellation.id,
                         zkb_icon='<img src="/static/sovtimer/images/zkillboard.png">',
                     )
                 )
-                defender_score_percent += constellation_killboard_link
+
+                campaign_progress_html += (
+                    constellation_killboard_link + campaign_progress_icon
+                )
 
             data.append(
                 {
-                    "campaign_id": campaign.campaign_id,
+                    # type column
                     "event_type": MAP_EVENT_TO_TYPE[campaign.event_type],
-                    "solar_system_id": campaign.solar_system_id,
+                    # system column + filter
                     "solar_system_name": solar_system_name,
                     "solar_system_name_html": solar_system_name_html,
-                    "constellation_id": campaign.constellation_id,
+                    # constellazion column + filter
                     "constellation_name": constellation_name,
                     "constellation_name_html": constellation_name_html,
-                    "structure_id": campaign.structure_id,
+                    # region column + filter
                     "region_name": region_name,
                     "region_name_html": region_name_html,
-                    "attackers_score": attackers_score_percent,
-                    "defender_id": campaign.defender_id,
+                    # defender column + filter
                     "defender_name": defender_name,
                     "defender_name_html": defender_name_html,
-                    "defender_score": defender_score_percent,
+                    # adm column
                     "adm": structure_adm,
+                    # start column
                     "start_time": start_time,
+                    # remaining column
                     "remaining_time": "",
                     "remaining_time_in_seconds": remaining_time_in_seconds,
+                    # progress column
+                    "campaign_progress": campaign_progress_html,
+                    # acive filter column (hidden)
                     "active_campaign": active_campaign,
                 }
             )